+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 * Now possible to define custom input/output value objects for the query builder.
 * Added a `EventStream` response sender to simplify sending [sever-sent events](https://developer.mozilla.org/en-US/docs/Web/API/Server-sent_events/Using_server-sent_events).
 * Permission calculations now fully support special bits (sticky, setgid and setuid).
+* Added the following methods to the `Permissions` class:
+ - `add`
+ - `remove`
 
 #### Deprecations